requirements.txt
bokeh==3.3.0
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.0
numba==0.58.1
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pynndescent==0.5.10
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
tbb==2021.10.0
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.1
tzdata==2023.3
umap-learn==0.5.4
xyzservices==2023.10.1
